Add permission checks

+++ app/Http/Controllers/PermissionController.php
@@ -3,6 +3,8 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Session;
 use Inertia\Inertia;
 use Spatie\Permission\Models\Permission;
 
@@ -10,6 +12,10 @@ class PermissionController extends Controller
 {
     public function index()
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $permissions = Permission::all();
         return Inertia::render('Admin/Permissions/Index', [
             'permissions' => $permissions
@@ -18,11 +24,19 @@ class PermissionController extends Controller
 
     public function create()
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         return Inertia::render('Admin/Permissions/Create');
     }
 
     public function store(Request $request)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:permissions,name',
         ]);
@@ -36,6 +50,10 @@ class PermissionController extends Controller
 
     public function edit(Permission $permission)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         return Inertia::render('Admin/Permissions/Edit', [
             'permission' => $permission
         ]);
@@ -43,6 +61,10 @@ class PermissionController extends Controller
 
     public function update(Request $request, Permission $permission)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $validated = $request->validate([
             'name' => 'required|string|max:255|unique:permissions,name,' . $permission->id,
         ]);
@@ -56,6 +78,10 @@ class PermissionController extends Controller
 
     public function destroy(Permission $permission)
     {
+        if (!Auth::user()->can('user-roles-all')) {
+            Session::flash('error', 'Permissão Negada!');
+            return redirect()->back();
+        }
         $permission->delete();
         return redirect()->route('permissions.index')->with('success', 'Permissão excluída.');
     }
